requirements.txt
asgiref==3.3.0
Django==3.1.3
django-smart-selects==1.5.9
gunicorn==20.0.4
littlenv==0.1.7
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
